requirements.txt
numpy==1.19.5
pandas==1.3.1  
tensorflow-gpu==2.5.0
tensorflow-datasets==4.4.0
tensorflow-hub==0.12.0  
giotto-tda==0.5.1
keras==2.7.0 
keras-flops==0.1.2
keras-tcn==3.3.0
matplotlib==3.5.0
jupyterlab==3.0.16
torch==1.10.0             
torchvision==0.11.1
tqdm==4.61.2
pytorch2keras==0.2.4 
onnx==1.8.0              
onnx-tf==1.8.0
onnx2keras==0.0.24
librosa==0.8.1
kapre==0.2.0
scikit-learn==0.24.2             
scipy==1.7.0 
arm-mango==1.1.0 

